% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{filter}
\alias{filter}
\title{Filters architectures without at least an average of barcode_threshold barcodes per batch with DNA counts > dna_threshold and RNA counts > rna_threshold.}
\usage{
filter(keju, dna_threshold = 5, rna_threshold = 5, barcode_threshold = 10)
}
\arguments{
\item{keju}{a keju object containing counts}

\item{dna_threshold}{minimum DNA counts per barcode}

\item{rna_threshold}{minimum RNA counts per barcode}

\item{barcode_threshold}{minimum average barcodes per batch}
}
\value{
a keju object with the filtered_counts object
}
\description{
Filters architectures without at least an average of barcode_threshold barcodes per batch with DNA counts > dna_threshold and RNA counts > rna_threshold.
}
