% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keju-package.R
\docType{package}
\name{keju-package}
\alias{keju}
\alias{keju-package}
\title{keju: Analyze Massively Parallel Reporter Assay Data}
\description{
A Bayesian hierarchical model for estimating transcription rate and differential activity in Massively Parallel Reporter Assay data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pimentellab/keju}
  \item \url{http://pimentellab.com/keju/}
  \item Report bugs at \url{https://github.com/pimentellab/keju/issues}
}

}
\author{
\strong{Maintainer}: Albert Xue \email{albertsxue@gmail.com}

}
\keyword{internal}
