% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keju.R
\name{keju_fit}
\alias{keju_fit}
\title{Fits keju model to data.}
\usage{
keju_fit(
  keju,
  output_dir,
  model = "no_motif",
  infer_differential_activity = FALSE,
  seed = 1,
  chains = 4,
  parallel_chains = 4
)
}
\arguments{
\item{keju}{a filtered and processed keju object}

\item{output_dir}{directory for output files}

\item{model}{model type. Must be one of 'no_motif', 'motif_shrinkage', or 'covariate_motif_slope_intercept.}

\item{infer_differential_activity}{boolean to use effect size estimation with a control and alternate treatment}

\item{seed}{random seed for MCMC sampling. See https://mc-stan.org/cmdstanr/reference/model-method-sample.html}

\item{chains}{number of Markov chains to use.  See https://mc-stan.org/cmdstanr/reference/model-method-sample.html}

\item{parallel_chains}{maximum number of MCMC chains to run in parallel. See https://mc-stan.org/cmdstanr/reference/model-method-sample.html}
}
\value{
a fitted keju object
}
\description{
Fits keju model to data.
}
