% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{keju_from_counts}
\alias{keju_from_counts}
\title{Create keju object from count data. Counts and metadata should be matched by barcode and batch.}
\usage{
keju_from_counts(
  barcode,
  R,
  D,
  batch,
  dna_batch,
  architecture,
  treatment,
  is_control_treatment = FALSE,
  is_control_architecture = FALSE,
  covariates = "correction",
  motif = NULL
)
}
\arguments{
\item{barcode}{a length N vector of barcodes}

\item{R}{a length N vector of RNA counts}

\item{D}{a length N vector of DNA counts}

\item{batch}{a length N vector denoting RNA batches}

\item{dna_batch}{a length N vector denoting DNA batches}

\item{architecture}{a length N vector of enhancer architectures or names. Transcription rate and effect size estimates will be fit on these.}

\item{treatment}{a length N vector detailing the treatments.}

\item{is_control_treatment}{a length N boolean vector denoting the control treatment.}

\item{is_control_architecture}{a length N boolean vector denoting architectures that are negative controls.}

\item{covariates}{a length N vector denoting covariates to be used in effect size correction and potentially slope/intercept fitting. If none, will default to a vector of ones of length N, to perform global effect size correction.}

\item{motif}{a length N vector denoting motifs shared across architectures. Used to shrink architecture-level estimates to a shared motif-level estimate. Necessary for motif_shrinkage and covariate_motif_slope_intercept models.}
}
\value{
an initialized keju object with a counts object
}
\description{
Create keju object from count data. Counts and metadata should be matched by barcode and batch.
}
