% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keju.R
\name{pretty_summarize}
\alias{pretty_summarize}
\title{Processes parameter estimates using existing metadata from the keju object. Calls significance for effect sizes and writes to output folder.}
\usage{
pretty_summarize(keju, model, output_folder, infer_differential_activity)
}
\arguments{
\item{keju}{a fitted keju object}

\item{model}{the type of model fit by keju}

\item{output_folder}{folder for output estimate csv files}

\item{infer_differential_activity}{boolean to use effect size estimation with a control and alternate treatment}
}
\description{
Processes parameter estimates using existing metadata from the keju object. Calls significance for effect sizes and writes to output folder.
}
