% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{use_covariate_slope_intercept}
\alias{use_covariate_slope_intercept}
\title{Process data to fit keju model with motif-level shrinkage and covariate-level effects on transcription rate.}
\usage{
use_covariate_slope_intercept(
  keju,
  G = 50,
  infer_differential_activity = FALSE
)
}
\arguments{
\item{keju}{a keju object containing counts or filtered_counts object.}

\item{G}{number of enhancers per dispersion parameter}

\item{infer_differential_activity}{boolean to use effect size estimation with a control and alternate treatment}
}
\value{
a keju object processed for the covariate_motif_slope_intercept model
}
\description{
Process data to fit keju model with motif-level shrinkage and covariate-level effects on transcription rate.
}
