% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{use_motif_shrinkage}
\alias{use_motif_shrinkage}
\title{Process counts to fit keju model with motif-level shrinkage.}
\usage{
use_motif_shrinkage(keju, G = 50, infer_differential_activity = FALSE)
}
\arguments{
\item{keju}{a keju object containing counts or filtered_counts object.}

\item{G}{number of enhancers per dispersion parameter}

\item{infer_differential_activity}{boolean to use effect size estimation with a control and alternate treatment}
}
\value{
a keju object processed for the motif_shrinkage model
}
\description{
Process counts to fit keju model with motif-level shrinkage.
}
