% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{use_no_motif}
\alias{use_no_motif}
\title{Process data to fit keju model without motif-level shrinkage. Prefers to run on filtered_counts, but will run on raw counts.}
\usage{
use_no_motif(keju, G = 50, infer_differential_activity = FALSE)
}
\arguments{
\item{keju}{a keju object containing counts}

\item{G}{number of enhancers per dispersion parameter}

\item{infer_differential_activity}{boolean to use effect size estimation with a control and alternate treatment}
}
\value{
a keju object processed for the no_motif model
}
\description{
Process data to fit keju model without motif-level shrinkage. Prefers to run on filtered_counts, but will run on raw counts.
}
